Begin hooking up controller
