+++ .gitignore
@@ -13,6 +13,7 @@ yarn-error.log
 coverage/
 *.sw*
 .env
+.nvm/
 
 # Database downloads
 latest.dump
